Refactor package name of commons module.
