+++ commons/src/test/java/org/lfenergy/compas/core/commons/ResourcesTest.java
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2021 Alliander N.V.
 //
 // SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.commons;
+package org.lfenergy.compas.core.commons;
 
 import org.junit.jupiter.api.Test;
