+++ commons/src/test/java/org/lfenergy/compas/core/commons/exception/CompasExceptionTest.java
@@ -1,12 +1,12 @@
 // SPDX-FileCopyrightText: 2021 Alliander N.V.
 //
 // SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.commons.exception;
+package org.lfenergy.compas.core.commons.exception;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.lfenergy.compas.commons.exception.CompasErrorCode.NO_CLASS_LOADER_ERROR_CODE;
+import static org.lfenergy.compas.core.commons.exception.CompasErrorCode.NO_CLASS_LOADER_ERROR_CODE;
 
 class CompasExceptionTest {
     @Test
