+++ commons/src/test/java/org/lfenergy/compas/core/commons/model/XmlElementBPojo.java
@@ -1,17 +1,17 @@
 // SPDX-FileCopyrightText: 2021 Alliander N.V.
 //
 // SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.commons.model;
+package org.lfenergy.compas.core.commons.model;
 
 import javax.xml.bind.annotation.*;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlRootElement(name = "XmlElementAPojo")
-public class XmlElementAPojo {
+@XmlType(name = "XmlElementBPojo")
+public class XmlElementBPojo {
     @XmlAttribute(name = "id", required = true)
     public String id;
-    @XmlElement(name = "Other")
-    public XmlElementBPojo other;
+    @XmlElement(name = "Name")
+    public String name;
 
     public String getId() {
         return id;
@@ -21,11 +21,11 @@ public class XmlElementAPojo {
         this.id = id;
     }
 
-    public XmlElementBPojo getOther() {
-        return other;
+    public String getName() {
+        return name;
     }
 
-    public void setOther(XmlElementBPojo other) {
-        this.other = other;
+    public void setName(String name) {
+        this.name = name;
     }
 }
