+++ scl2003/src/main/java/org/lfenergy/compas/scl2003/commons/Scl2003MarshallerWrapper.java
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.lfenergy.compas.scl2003.commons;
 
-import org.lfenergy.compas.commons.MarshallerWrapper;
+import org.lfenergy.compas.core.commons.MarshallerWrapper;
 import org.lfenergy.compas.scl2003.model.SCL;
 
 import javax.xml.bind.Marshaller;
