+++ scl2003/src/test/java/org/lfenergy/compas/scl2003/commons/Scl2003MarshallerWrapperTest.java
@@ -4,7 +4,7 @@
 package org.lfenergy.compas.scl2003.commons;
 
 import org.junit.jupiter.api.Test;
-import org.lfenergy.compas.commons.exception.CompasException;
+import org.lfenergy.compas.core.commons.exception.CompasException;
 import org.lfenergy.compas.scl.extensions.commons.CompasExtensionsConstants;
 import org.lfenergy.compas.scl.extensions.model.ObjectFactory;
 import org.lfenergy.compas.scl2003.model.SCL;
@@ -14,7 +14,7 @@ import org.lfenergy.compas.scl2003.model.TPrivate;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.lfenergy.compas.commons.exception.CompasErrorCode.UNMARSHAL_ERROR_CODE;
+import static org.lfenergy.compas.core.commons.exception.CompasErrorCode.UNMARSHAL_ERROR_CODE;
 
 class Scl2003MarshallerWrapperTest {
     @Test
