+++ scl2007b/src/main/java/org/lfenergy/compas/scl2007b/commons/Scl2007bMarshallerWrapper.java
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.lfenergy.compas.scl2007b.commons;
 
-import org.lfenergy.compas.commons.MarshallerWrapper;
+import org.lfenergy.compas.core.commons.MarshallerWrapper;
 import org.lfenergy.compas.scl2007b.model.SCL;
 
 import javax.xml.bind.Marshaller;
